requirements.txt
streamlit
groq
tavily-python
python-dotenv
requests

langchain==0.1.16
langchain-core==0.1.45
langchain-community==0.0.38
langchain-groq
